+++ src/app/dataset-view/additional-components/dataset-settings-component/dataset-settings.helpers.ts
@@ -31,6 +31,7 @@ export function isSettingsTabAccessibleHelper(
             );
 
         case SettingsTabsEnum.SCHEDULING:
+        case SettingsTabsEnum.INGEST_CONFIGURATION:
             return (
                 featureFlags.enableScheduling &&
                 datasetBasics.kind === DatasetKind.Root &&