+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/ingest-configuration/ingest-configuration-form/ingest-configuration.module.ts
@@ -10,10 +10,11 @@ import { CommonModule } from "@angular/common";
 import { IngestConfigurationFormComponent } from "./ingest-configuration-form.component";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { MatCheckboxModule } from "@angular/material/checkbox";
+import { MatDividerModule } from "@angular/material/divider";
 
 @NgModule({
     declarations: [IngestConfigurationFormComponent],
     exports: [IngestConfigurationFormComponent],
-    imports: [CommonModule, FormsModule, MatCheckboxModule, ReactiveFormsModule],
+    imports: [CommonModule, FormsModule, MatDividerModule, MatCheckboxModule, ReactiveFormsModule],
 })
 export class IngestConfigurationModule {}